requirements.txt
discord.py==2.3.2
wavelink==3.2.0
spotipy==2.24.0
aiohttp==3.10.5
